Add search to post pages
